CIR-2006 platops-pr-bot-suggestions - multiple routes files using the same package
