+++ conf/appV2.routes
@@ -1,3 +1,3 @@
 # Stripped down api only endpoints
-POST        /send-code                              uk.gov.hmrc.emailverification.controllers.EmailVerificationV2Controller.sendCode()
-POST        /verify-code                            uk.gov.hmrc.emailverification.controllers.EmailVerificationV2Controller.verifyCode()
+POST        /send-code                              uk.gov.hmrc.emailverification.controllers.v2.EmailVerificationV2Controller.sendCode()
+POST        /verify-code                            uk.gov.hmrc.emailverification.controllers.v2.EmailVerificationV2Controller.verifyCode()
