+++ conf/application.conf
@@ -74,7 +74,7 @@ controllers {
     needsAuditing = true
   }
 
-  uk.gov.hmrc.emailverification.controllers.EmailVerificationV2Controller = {
+  uk.gov.hmrc.emailverification.controllers.v2.EmailVerificationV2Controller = {
     needsLogging = true
     needsAuditing = true
   }
