+++ conf/testOnlyDoNotUseInAppConf.routes
@@ -10,8 +10,5 @@
 # Failing to follow this rule may result in test routes deployed in production.
 
 # Add all the application routes to the prod.routes file
-->          /                                            prod.Routes
-
-GET         /test-only/passcodes                         uk.gov.hmrc.emailverification.controllers.EmailPasscodeController.testOnlyGetPasscodes()
-
-POST        /test-only/retrieve/verification-code        uk.gov.hmrc.emailverification.controllers.TestOnlyEmailVerificationV2Controller.getVerificationCode
+->        /test-only           test.Routes
+->        /                    prod.Routes
